+++ TrafficMAS/src/nl/uu/trafficmas/agent/PregnantWomanAgent.java
@@ -14,7 +14,7 @@ public class PregnantWomanAgent extends Agent {
 	}
 
 	@Override
-	public double specificUtility(int arrivalTime, ArrayList<Sanction> sanctionList) {
+	public double specificUtility(double arrivalTime, ArrayList<Sanction> sanctionList) {
 		double utility = 0;
 		utility = (double) (this.getGoalArrivalTime() / (3*arrivalTime));
